move `load_loop_order` to a separate utils file
