+++ tests/model_connectors/test_district_single_ghe.py
@@ -1,11 +1,11 @@
 # :copyright (c) URBANopt, Alliance for Sustainable Energy, LLC, and other contributors.
 # See also https://github.com/urbanopt/geojson-modelica-translator/blob/develop/LICENSE.md
 
-import json
 from pathlib import Path
 
 import pytest
 
+from geojson_modelica_translator.external_package_utils import load_loop_order
 from geojson_modelica_translator.geojson.urbanopt_geojson import UrbanOptGeoJson
 from geojson_modelica_translator.model_connectors.couplings.coupling import Coupling
 from geojson_modelica_translator.model_connectors.couplings.graph import CouplingGraph
@@ -28,12 +28,15 @@ class DistrictSystemTest(TestCaseBase):
         self.data_dir, self.output_dir = self.set_up(Path(__file__).parent, project_name)
 
         # load in the example geojson with multiple buildings
-        filename = Path(self.data_dir) / "time_series_ex2.json"
-        self.gj = UrbanOptGeoJson(filename)
+        geojson_filename = Path(self.data_dir) / "time_series_ex2.json"
+        self.gj = UrbanOptGeoJson(geojson_filename)
 
         # load system parameter data
-        filename = Path(self.data_dir) / "system_params_ghe_3.json"
-        sys_params = SystemParameters(filename)
+        sys_param_filename = Path(self.data_dir) / "system_params_ghe_3.json"
+        sys_params = SystemParameters(sys_param_filename)
+
+        # read the loop order and create building groups
+        loop_order: list = load_loop_order(sys_param_filename)
 
         # create ambient water loop stub
         ambient_water_stub = NetworkDistributionPump(sys_params)
@@ -43,11 +46,6 @@ class DistrictSystemTest(TestCaseBase):
 
         # create district data
         design_data = DesignDataSeries(sys_params)
-
-        # read the loop order and create building groups
-        filename = Path(self.data_dir) / "_loop_order.json"
-        loop_order: list = json.loads(filename.read_text())
-        print(filename)
 
         # create the couplings and graph
         all_couplings = []
